+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Mobile Center SDK for iOS Change Log
 
+## Next version
+
+This version has bug fixes.
+
+### MobileCenter
+
+* **[Bug]** Fix http tasks cancelled when expected to be suspended.
+
+___
 
 ## Version 0.3.6