+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This version has bug fixes.
 
 ### MobileCenter
 
+* **[Bug]** Fix crash sending failure callback, http status code now included in forwarded error.
 * **[Bug]** Fix http tasks cancelled when expected to be suspended.
 
 ___